+++ Source/UserSession/ZMUserSession+Background.m
@@ -333,7 +333,7 @@ static NSString *ZMLogTag = @"Push";
         BOOL callIsStillOngoing = (note.conversation.callParticipants.count > 0) || note.conversation.voiceChannel.state == ZMVoiceChannelStateIncomingCall; // FIXME check if this logic works
         BOOL userWantsToCallBack = ([note.category isEqualToString:ZMMissedCallCategory]);
         
-        if ([WireCallCenterT activeCallConversationsInUserSession:self].count == 0 && (callIsStillOngoing || userWantsToCallBack)) {
+        if ([WireCallCenter activeCallConversationsInUserSession:self].count == 0 && (callIsStillOngoing || userWantsToCallBack)) {
             [note.conversation.voiceChannel joinWithVideo:NO];
             [note.conversation.managedObjectContext saveOrRollback];
         }
